Several minor fixes and improvements. Time delay fields changed from 8-bit to 32-bit. Tidied up URL for settings page. Extended sensors page from 2 to 3 one-wire temperature sensors. Minor change to report format. Added debug output of raw temperature data. Improved and corrected floating point number formatting.

+++ src/globals.h
@@ -30,8 +30,8 @@ extern char *p_onewireids;
 
 struct PERSISTENT_DATA {    // this structure can be stored in EEPROM
     uint16_t port;
-    uint8_t time_to_sleep;
-    uint8_t time_between_reports;
+    uint32_t time_to_sleep;
+    uint32_t time_between_reports;
     uint8_t dht22_pin;
     uint8_t onewire_pin;
     int8_t  rot;
